% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootFA.R
\name{bootFA}
\alias{bootFA}
\title{Nonparametric bootstrap to evaluate the stability of the internal structure}
\usage{
bootFA(
  fit,
  R = 100,
  n.cores = 2,
  plots = TRUE,
  rm.files = TRUE,
  seed = NULL,
  verbose = TRUE,
  digits = 3
)
}
\arguments{
\item{fit}{A \code{wrapFA} object or a \code{list} of \code{wrapFA} objects.}

\item{R}{A \code{double} indicating the number of bootstrapping replications. Default is 100.}

\item{n.cores}{A \code{double} indicating the number of CPU processors to be used for the bootstrapping. Default is 2.}

\item{plots}{A \code{logical} indicating whether result plots should be generated. Default is \code{TRUE}.}

\item{rm.files}{A \code{logical} indicating whether the stored files should be deleted. The default is \code{TRUE}.}

\item{seed}{A \code{double} containing a random seed. Default is \code{NULL}.}

\item{verbose}{A \code{logical} indicating whether a progress bar should be printed. Default is \code{TRUE}.}
}
\value{
\code{bootFA} returns an object of class \code{bootFA}.
\describe{
\item{\code{lambdaCC}}{Average congruent coefficient of factor loadings between pairs of bootraspping solutions (\code{matrix}).}
\item{\code{lambdaMAD}}{Average mean absolute deviation of factor loadings  between pairs of bootraspping solutions (\code{matrix}).}
\item{\code{phiCC}}{Average congruent coefficient of factor correlations between pairs of bootraspping solutions (\code{matrix}).}
\item{\code{phiMAD}}{Average mean absolute deviation of factor correlations between pairs of bootraspping solutions (\code{matrix}).}
\item{\code{lambdaCC.r}}{Congruent coefficient of factor loadings between pairs of bootraspping solutions (\code{list}).}
\item{\code{lambdaMAD.r}}{Mean absolute deviation of factor loadings  between pairs of bootraspping solutions (\code{list}).}
\item{\code{phiCC.r}}{Congruent coefficient of factor correlations between pairs of bootraspping solutions (\code{list}).}
\item{\code{phiMAD.r}}{Mean absolute deviation of factor correlations between pairs of bootraspping solutions (\code{list}).}
\item{\code{plots}}{Plots summarizing the results. Only printed if \code{plots = TRUE} (\code{list}).}
\item{\code{specifications}}{Specifications used to run the function (\code{list}).}
}
}
\description{
Evaluates the stability of a factor analysis solution (e.g., CFA, EFA, ECFA, BSEM) by conducting nonparametric
bootstrapping, resampling with a replacement from the original dataset. This is used in Nájera et al. (2022), based on the
bootstrap EGA (Christensen & Golino, 2021).
}
\examples{
library(MBESS)
data(HS)
HS <- HS[, -c(1:8)]
colnames(HS) <- paste0("t", 1:26)
modHS <- data.frame(spatial = c(rep(1, 4), rep(0, 20), rep(1, 2)),
                    verbal = c(rep(0, 4), rep(1, 5), rep(0, 17)),
                    speed = c(rep(0, 9), rep(1, 4), rep(0, 13)),
                    memory = c(rep(0, 13), rep(1, 6), rep(0, 7)),
                    maths = c(rep(0, 19), rep(1, 5), rep(0, 2)), row.names = colnames(HS))
modHS.lav <- modFA(lambda = modHS, software = "lavaan", keep.names = TRUE)
cfa.lav <- CFA(model = modHS.lav$CFA, data = HS)
boot.cfa <- bootFA(fit = cfa.lav, n.cores = 4)
}
\references{
Christensen, A. P., & Golino, H. (2021). Estimating the stability of the number of factors via bootstrap exploratory graph analysis: A tutorial. \emph{Psych}, \emph{3}(3), 479-500. https://doi.org/10.3390/psych3030032
Nájera, P., Abad, F. J., & Sorrel, M. A. (2022). Is EFA always to be preferred? A systematic comparison of factor analytic techniques throughout the confirmatory-exploratory continuum. \emph{Manuscript submitted for publication}.
}
