% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFA.R
\name{CFA}
\alias{CFA}
\title{Confirmatory factor analysis (CFA)}
\usage{
CFA(
  model,
  data,
  categorical = NULL,
  estimator = NULL,
  software = NULL,
  mimic.mplus = TRUE,
  mplus.path = NULL,
  rm.files = FALSE,
  max.iter = 10000,
  mod = "none",
  mod.args = list(stepwise = TRUE, mod.error = TRUE, MI.cut = 3.841459, MI.power = 0.75,
    SEPC.lambda = 0.4, SEPC.psi = 0.1, max.mods = 50),
  suppressMessages = TRUE
)
}
\arguments{
\item{model}{Either a \code{character} string containing the user-specified CFA model using either the \emph{lavaan} or the \emph{Mplus} syntax, or a \code{list} containing the design Lambda \code{matrix} (i.e., factor loadings), Phi \code{matrix} (i.e., factor correlations), and Psi \code{matrix} (i.e., error correlations), in that specific order. In the design matrices, free and fixed parameters must be indicated with 1 and 0, respectively.}

\item{data}{A \emph{N} individuals x \emph{J} observed variables \code{matrix} or \code{data.frame}. Missing values need to be coded as \code{NA}.}

\item{categorical}{A \code{vector} containing the number or name of the observed variables that need to be treated as categorical. Default is \code{NULL}, which treates all observed variables as continuous.}

\item{estimator}{A \code{character} indicating the estimator to be used to fit the model. It can be \code{"ML"}, \code{"MLM"}, \code{"MLMV"}, \code{"MLMVS"}, \code{"MLF"}, \code{"MLR"}, \code{"GLS"}, \code{"DWLS"}, \code{"WLS"}, \code{"WLSM"}, \code{"WLSMV"}, \code{"ULS"},\code{"ULSM"}, and \code{"ULSMV"}. Check \code{?lavaan::lavOptions} or the \emph{Mplus User's Guide} for more information. Default is \code{NULL}, which uses \code{"ML"} if all observed variables are continuous and \code{"WLSMV"} if any observed variable is categorical.}

\item{software}{A \code{character} indicating the software to use for fitting the CFA model. It can be \code{"lavaan"} or \code{"mplus"}. Only required if design matrices are used in the \code{model} argument; the software is automatically detected if a syntax is used. Default is \code{NULL}.}

\item{mimic.mplus}{A \code{logical} indicating whether \emph{lavaan} should mimic \emph{Mplus} output. Only applicable to \emph{lavaan}. Default is \code{TRUE}.}

\item{mplus.path}{A \code{character} indicating the path where the the files generated by \emph{Mplus} will be stored, including the name of the files. Only applicable to \emph{Mplus}. Default is \code{NULL}, which stores the files in the current working directory with the name \emph{wrapCFA}.}

\item{rm.files}{A \code{logical} indicating whether the stored files should be deleted. The default is \code{FALSE}.}

\item{max.iter}{A \code{double} indicating the maximum number of iterations. Default is 10000.}

\item{mod}{A \code{character} indicating whether sequential model modifications are permitted. It can be \code{"none"} for no modifications, \code{"MI"} for modifications based solely on modification indices, or \code{"Saris"} for the Saris procedure (Saris et al., 2009). \code{"Saris"} is only applicable to \emph{lavaan}. Default is \code{"none"}.}

\item{mod.args}{A list of arguments for specifying the sequential model modification search:
\describe{
\item{\code{stepwise}}{A \code{logical} indicating whether model modifications should be included sequentially (rather than in a single step). Default is \code{TRUE}.}
\item{\code{mod.error}}{A \code{logical} indicating whether error correlations should be permitted. Default is \code{TRUE}.}
\item{\code{MI.cut}}{A \code{numeric} indicating the MI cutoff point. Default is \code{3.841459}.}
\item{\code{MI.power}}{A \code{numeric} indicating the cutoff for the power associated to the MI in the Saris procedure. Default is 0.75.}
\item{\code{SEPC.lambda}}{A \code{numeric} indicating the cutoff for the standardized expected parameter change (SEPC) associated to factor loadings in the Saris procedure. Default is 0.4.}
\item{\code{SEPC.psi}}{A \code{numeric} indicating the cutoff for the SEPC associated to error correlations in the Saris procedure. Default is 0.1.}
\item{\code{max.mods}}{A \code{double} indicating the maximum number of modifications permitted. Default is 50.}
}}

\item{suppressMessages}{A \code{logical} indicating whether the messages from MplusAutomation should be silent. Default is \code{TRUE}.}
}
\value{
\code{CFA} returns an object of class \code{wrapFA}.
\describe{
\item{\code{model.fit}}{Model fit indices (\code{vector}).}
\item{\code{lambda}}{The standardized estimated factor loading matrix (\code{matrix}).}
\item{\code{phi}}{The standardized estimated factor correlation matrix (\code{matrix}).}
\item{\code{psi}}{The standardized estimated error correlation matrix (\code{matrix}).}
\item{\code{lambda.p}}{The p-value associated to each free parameter in the estimated factor loading matrix (\code{matrix}).}
\item{\code{phi.p}}{The p-value associated to each free parameter in the estimated factor correlation matrix (\code{matrix}).}
\item{\code{psi.p}}{The p-value associated to each free parameter in the estimated error correlation matrix (\code{matrix}).}
\item{\code{MI}}{Information about the MI and SEPC of the fixed parameters (\code{matrix}).}
\item{\code{mods}}{The modifications included in the model (\code{matrix}).}
\item{\code{out}}{The complete output provided by either \emph{lavaan} or \emph{Mplus} (\code{list}).}
\item{\code{specifications}}{Specifications used to run the function (\code{list}).}
}
}
\description{
Fit a CFA model via \emph{lavaan} or \emph{MplusAutomation}. Sequential model modifications can be implemented
by using modification indices (MI) or the procedure proposed by Saris et al. (2009).
}
\examples{
\dontrun{
library(MBESS)
data(HS)
HS <- HS[, -c(1:8)]
colnames(HS) <- paste0("t", 1:26)
modHS <- data.frame(spatial = c(rep(1, 4), rep(0, 20), rep(1, 2)),
                    verbal = c(rep(0, 4), rep(1, 5), rep(0, 17)),
                    speed = c(rep(0, 9), rep(1, 4), rep(0, 13)),
                    memory = c(rep(0, 13), rep(1, 6), rep(0, 7)),
                    maths = c(rep(0, 19), rep(1, 5), rep(0, 2)), row.names = colnames(HS))
modHS.lav <- modFA(lambda = modHS, software = "lavaan", keep.names = TRUE)
cfa.lav <- CFA(model = modHS.lav$CFA, data = HS)
}
}
\references{
Saris, W. E., Satorra, A., & van der Veld, W. M. (2009). Testing structural equation models or detection of misspecifications?. \emph{Structural Equation Modeling}, \emph{16}, 561-582. https://doi.org/10.1080/10705510903203433
}
