% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFA.R
\name{EFA}
\alias{EFA}
\title{Exploratory factor analysis (EFA)}
\usage{
EFA(
  model = NULL,
  data,
  n.factors = NULL,
  categorical = NULL,
  estimator = NULL,
  rotation = "Geomin",
  software = NULL,
  mimic.mplus = TRUE,
  mplus.path = NULL,
  rm.files = FALSE,
  max.iter = 10000,
  mod = "none",
  mod.args = list(stepwise = TRUE, mod.error = TRUE, MI.cut = 3.841459, MI.power = 0.75,
    SEPC.lambda = 0.4, SEPC.psi = 0.1, max.mods = 50),
  suppressMessages = TRUE
)
}
\arguments{
\item{model}{A \code{character} string containing the EFA model using either the \emph{lavaan} or \emph{Mplus} syntax. It can be null if the number of factors is specified instead in the argument \code{n.factors}.}

\item{data}{A \emph{N} individuals x \emph{J} observed variables \code{matrix} or \code{data.frame}. Missing values need to be coded as \code{NA}.}

\item{n.factors}{A \code{double} indicating the number of factors to extract. It can be null if \code{model} is specified instead.}

\item{categorical}{A \code{vector} containing the number or name of the observed variables that need to be treated as categorical. Default is \code{NULL}, which treates all observed variables as continuous.}

\item{estimator}{A \code{character} indicating the estimator to be used to fit the model. It can be \code{"ML"}, \code{"MLM"}, \code{"MLMV"}, \code{"MLMVS"}, \code{"MLF"}, \code{"MLR"}, \code{"GLS"}, \code{"DWLS"}, \code{"WLS"}, \code{"WLSM"}, \code{"WLSMV"}, \code{"ULS"},\code{"ULSM"}, and \code{"ULSMV"}. Check \code{?lavaan::lavOptions} or the \emph{Mplus} User's Guide for more information. Default is \code{NULL}, which uses \code{"ML"} if all observed variables are continuous and \code{"WLSMV"} if any observed variable is categorical.}

\item{rotation}{A \code{character} indicating the rotation procedure to be used. It can be \code{"Geomin"}, \code{"Target"}, \code{"RETAM"}, \code{"Oblimin"}, \code{"Promax"}, \code{"Varimax"}, and \code{"Quartimin"}. \code{"Target"} and \code{"RETAM"} are only available for \emph{Mplus}.}

\item{software}{A \code{character} indicating the software to use for fitting the EFA model. It can be \code{"lavaan"} or \code{"mplus"}. Only required if design matrices are used in the \code{model} argument; the software is automatically detected if a syntax is used. Default is \code{NULL}.}

\item{mimic.mplus}{A \code{logical} indicating whether \emph{lavaan} should mimic \emph{Mplus} output. Only applicable to \emph{lavaan}. Default is \code{TRUE}.}

\item{mplus.path}{A \code{character} indicating the path where the the files generated by \emph{Mplus} will be stored, including the name of the files. Default is \code{NULL}, which stores the files in the current working directory with the name \emph{wrapEFA}.}

\item{rm.files}{A \code{logical} indicating whether the stored files should be deleted. The default is \code{FALSE}.}

\item{max.iter}{A \code{double} indicating the maximum number of iterations. Default is 10000.}

\item{mod}{A \code{character} indicating whether sequential model modifications are permitted. It can be \code{"none"} for no modifications, \code{"MI"} for modifications based solely on modification indices, or \code{"Saris"} for the Saris procedure (Saris et al., 2009). \code{"Saris"} is only applicable to \emph{lavaan}. Default is \code{"none"}.}

\item{mod.args}{A \code{list} of arguments for specifying the sequential model modification search:
\describe{
\item{\code{stepwise}}{A \code{logical} indicating whether model modifications should be included sequentially (rather than in a single step). Default is \code{TRUE}.}
\item{\code{mod.error}}{A \code{logical} indicating whether error correlations should be permitted. Default is \code{TRUE}.}
\item{\code{MI.cut}}{A \code{numeric} indicating the MI cutoff point. Default is \code{3.841459}.}
\item{\code{MI.power}}{A \code{numeric} indicating the cutoff for the power associated to the MI in the Saris procedure. Default is 0.75.}
\item{\code{SEPC.lambda}}{A \code{numeric} indicating the cutoff for the standardized expected parameter change (SEPC) associated to factor loadings in the Saris procedure. Default is 0.4.}
\item{\code{SEPC.psi}}{A \code{numeric} indicating the cutoff for the SEPC associated to error correlations in the Saris procedure. Default is 0.1.}
\item{\code{max.mods}}{A \code{double} indicating the maximum number of modifications permitted. Default is 50.}
}}

\item{suppressMessages}{A \code{logical} indicating whether the messages from MplusAutomation should be silent. Default is \code{TRUE}.}
}
\value{
\code{EFA} returns an object of class \code{wrapFA}.
\describe{
\item{\code{model.fit}}{Model fit indices (\code{vector}).}
\item{\code{lambda}}{The standardized estimated factor loading matrix (\code{matrix}).}
\item{\code{phi}}{The standardized estimated factor correlation matrix (\code{matrix}).}
\item{\code{psi}}{The standardized estimated error correlation matrix (\code{matrix}).}
\item{\code{lambda.p}}{The p-value associated to each free parameter in the estimated factor loading matrix (\code{matrix}).}
\item{\code{phi.p}}{The p-value associated to each free parameter in the estimated factor correlation matrix (\code{matrix}).}
\item{\code{psi.p}}{The p-value associated to each free parameter in the estimated error correlation matrix (\code{matrix}).}
\item{\code{MI}}{Information about the MI and SEPC of the fixed parameters (\code{matrix}).}
\item{\code{mods}}{The modifications included in the model (\code{matrix}).}
\item{\code{out}}{The complete output provided by either \emph{lavaan} or \emph{Mplus} (\code{list}).}
\item{\code{specifications}}{Specifications used to run the function (\code{list}).}
}
}
\description{
Fit an EFA model via \emph{lavaan} or \emph{MplusAutomation}. Correlated residuals can be included in the
model as in exploratory structural equation modeling (ESEM; Asparouhov & Muthén, 2009). Several rotation procedures are supported. The objectively
refined target matrix (RETAM) procedure (Lorenzo-Seva & Ferrando, 2020) is also available if \emph{MplusAutomation} is used. Moreover, sequential
model modifications can be implemented by using modification indices (MI) or the procedure proposed by Saris et al. (2009).
}
\examples{
\donttest{
library(MBESS)
data(HS)
HS <- HS[, -c(1:8)]
colnames(HS) <- paste0("t", 1:26)
modHS <- data.frame(spatial = c(rep(1, 4), rep(0, 20), rep(1, 2)),
                    verbal = c(rep(0, 4), rep(1, 5), rep(0, 17)),
                    speed = c(rep(0, 9), rep(1, 4), rep(0, 13)),
                    memory = c(rep(0, 13), rep(1, 6), rep(0, 7)),
                    maths = c(rep(0, 19), rep(1, 5), rep(0, 2)), row.names = colnames(HS))
modHS.lav <- modFA(lambda = modHS, software = "lavaan", keep.names = TRUE)
modHS.mpl <- modFA(lambda = modHS, software = "mplus", keep.names = TRUE)
efa_geomin <- EFA(model = modHS.lav$EFA, data = HS)
efa_target <- EFA(model = modHS.mpl$EFAt, data = HS, rotation = "TARGET")
}
}
\references{
Asparouhov, T., & Muthén, B. O. (2009). Exploratory structural equation modeling. \emph{Structural Equation Modeling: A Multidisciplinary Journal}, \emph{16}(3), 397-438. https://doi.org/10.1080/10705510903008204
Lorenzo-Seva, U., & Ferrando, P. J. (2020). Unrestricted factor analysis of multidimensional test items based on an objectively refined target matrix. \emph{Behavior Research Methods}, \emph{52}, 116-130. https://doi.org/10.3758/s13428-019-01209-1
Saris, W. E., Satorra, A., & van der Veld, W. M. (2009). Testing structural equation models or detection of misspecifications?. \emph{Structural Equation Modeling}, \emph{16}, 561-582. https://doi.org/10.1080/10705510903203433
}
