% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modFA.R
\name{modFA}
\alias{modFA}
\title{Translate design matrices into syntax for lavaan or Mplus models}
\usage{
modFA(lambda, phi = NULL, psi = NULL, software, keep.names = TRUE)
}
\arguments{
\item{lambda}{Factor loading \code{matrix} (or \code{data.frame}).}

\item{phi}{Factor correlation \code{matrix} (or \code{data.frame}). Default is \code{NULL}, which implies that all factor correlations are included in the model.}

\item{psi}{Residual correlation \code{matrix} (or \code{data.frame}). Default is \code{NULL}, which implies that no residual correlations are included in the model.}

\item{software}{A \code{character} indicating the software to use for translating the design matrices into syntax. It can be \code{"lavaan"}, \code{"mplus"}, or \code{"lslx"}.}

\item{keep.names}{A \code{logical} indicating whether the names of the observed and latent variables (taken from \code{lambda}) should be retained or not. Default is \code{TRUE}. If \code{FALSE}, \code{x1, x2, ...} is used for observed variables and \code{F1, F2, ...} for latent variables.}
}
\value{
\code{modFA} returns an object of class \code{modFA}.
\describe{
\item{\code{CFA}}{The syntax for fitting a CFA (\code{character}).}
\item{\code{EFA}}{The syntax for fitting an EFA with mechanical/analytical rotation (\code{character}).}
\item{\code{EFAt}}{The syntax for fitting an EFA with target rotation. Only available for \emph{Mplus} (\code{character}).}
\item{\code{BSEM}}{The syntax for fitting a BSEM. Only available for \emph{Mplus} (\code{character}).}
\item{\code{specifications}}{Specifications used to implement the \code{modFA} function (\code{list}).}
}
}
\description{
Translates design matrices (factor loadings, factor correlations, residual correlations) into a syntax that is compatible with either
\emph{lavaan} and \emph{Mplus}. Namely, CFA and EFA are supported for \emph{lavaan}, and CFA, EFA, EFA with target rotation, and BSEM are supported for \emph{Mplus}.
}
\examples{
bfi <- psych::bfi
bfi <- bfi[,1:25]
modelbfi <- data.frame(agree = c(rep(1, 5), rep(0, 20)),
                       consc = c(rep(0, 5), rep(1, 5), rep(0, 15)),
                       extra = c(rep(0, 10), rep(1, 5), rep(0, 10)),
                       neuro = c(rep(0, 15), rep(1, 5), rep(0, 5)),
                       openn = c(rep(0, 20), rep(1, 5)), row.names = colnames(bfi))
modelbfi.lav <- modFA(lambda = modelbfi, software = "lavaan", keep.names = TRUE)
modelbfi.mpl <- modFA(lambda = modelbfi, software = "mplus", keep.names = TRUE)
}
